refactor(ios): switch google sign in to supabase oauth web flow
GoogleSignIn-iOS 7.1 uses AppAuth which silently injects a nonce
into the OAuth request and never exposes it. Google echoes it back
in the id_token's nonce claim, so Supabase rejects the exchange
("Passed nonce and nonce in id_token should either both exist or
not."). 9.x adds a public nonce param but its SwiftPM manifest is
broken (invalid GoogleUtilities path).

Switch to client.auth.signInWithOAuth(provider: .google) which
opens an ASWebAuthenticationSession that intercepts the
pebbles://auth-callback redirect. Reuses the Google web client
already configured in Supabase, no native SDK needed. Apple stays
native — its flow controls the nonce explicitly via
ASAuthorizationController and is unaffected.

Removes GoogleSignIn-iOS dependency, GoogleSignInService,
GoogleIOSClientID/URL scheme wiring, and GIDSignIn config in
PebblesApp. Extracts nonce helpers to OAuthNonce.swift so the
remaining Apple flow uses them cleanly.

Requires "pebbles://auth-callback" in Supabase dashboard
Authentication → URL Configuration → Redirect URLs.
