+++ apps/ios/Config/Secrets.example.xcconfig
@@ -6,8 +6,3 @@
 
 SUPABASE_URL = https:/$()/your-project.supabase.co
 SUPABASE_ANON_KEY =
-
-// Google iOS OAuth client (from Google Cloud Console).
-// REVERSED form is used as a custom URL scheme — keep both in sync.
-GOOGLE_IOS_CLIENT_ID =
-GOOGLE_IOS_REVERSED_CLIENT_ID =
