+++ apps/ios/Pebbles/Resources/Info.plist
@@ -32,17 +32,6 @@
 	<string>$(SUPABASE_URL)</string>
 	<key>SupabaseAnonKey</key>
 	<string>$(SUPABASE_ANON_KEY)</string>
-	<key>GoogleIOSClientID</key>
-	<string>$(GOOGLE_IOS_CLIENT_ID)</string>
-	<key>CFBundleURLTypes</key>
-	<array>
-		<dict>
-			<key>CFBundleURLSchemes</key>
-			<array>
-				<string>$(GOOGLE_IOS_REVERSED_CLIENT_ID)</string>
-			</array>
-		</dict>
-	</array>
 	<key>CFBundleLocalizations</key>
 	<array>
 		<string>en</string>
